added access denial to unconnected users for figure manipulation
