+++ config/packages/security.yaml
@@ -35,11 +35,8 @@ security:
             form_login:
                 enable_csrf: true
 
-    # Easy way to control access for large sections of your site
-    # Note: Only the *first* access control that matches will be used
     access_control:
-        # - { path: ^/admin, roles: ROLE_ADMIN }
-        # - { path: ^/profile, roles: ROLE_USER }
+        - { path: '^/snowtricks/figure/', roles: IS_AUTHENTICATED }
 
 when@test:
     security:
