+++ src/Controller/HomeController.php
@@ -31,7 +31,7 @@ class HomeController extends AbstractController
         ]);
     }
 
-    #[Route('/snowtricks/figure/show/{id}', name: 'app_figure_show')]
+    #[Route('/snowtricks/show/{id}', name: 'app_figure_show')]
     public function show(Figure $figure, CommentaireRepository $repo, Request $request,ManagerRegistry $doctrine)
     {        
         $commentaire = new Commentaire();
@@ -65,69 +65,5 @@ class HomeController extends AbstractController
             'commentaires' => $commentaires,
             'formCommentaire' =>$form->createView()
         ]);
-    }
-
-    /**
-     * Suppression des images
-     */
-    #[Route('/snowtricks/figure/{id}/image/delete', name: 'app_figure_image_delete',  methods:["DELETE", "GET"])]
-    public function deleteImage(Images $image,ManagerRegistry $doctrine)
-    {
-
-        $figureId = $image->getFigure()->getId();
-
-        $nomImg = $image->getNom();
-        // On unlink le fichier
-        unlink($this->getParameter('images_directory').'/'.$nomImg);
-
-        // On supprime l'entrée en db
-        $entityManager = $doctrine->getManager();
-        $entityManager->remove($image);
-        $entityManager->flush();
-
-        return $this->redirectToRoute('app_figure_edit', ['id' => $figureId]);
-    }
-
-    /**
-     * Suppression des videos
-     */
-    #[Route('/snowtricks/figure/{id}/video/delete', name: 'app_figure_video_delete', methods:["DELETE", "GET"])]
-    public function deleteVideo(Video $video,ManagerRegistry $doctrine)
-    {
-
-        $figureId = $video->getFigure()->getId();
-
-        // On supprime l'entrée en db
-        $entityManager = $doctrine->getManager();
-        $entityManager->remove($video);
-        $entityManager->flush();
-
-        return $this->redirectToRoute('app_figure_edit', ['id' => $figureId]);
-    }
-
-    /**
-     * Suppression de la figure
-     */
-    #[Route('/snowtricks/figure/{id}/delete', name: 'app_figure_delete')]
-    public function deleteFigure(Figure $figure, ManagerRegistry $doctrine)
-    {
-        $entityManager = $doctrine->getManager();
-
-        // Supression des images dans le file
-        // On cherche d'abord à récupérer les noms des images
-        $figureId = $figure->getId();
-        $repo = $entityManager->getRepository(Images::class);
-        $imagesNom = $repo->findNameByFigureId($figureId);
-
-        $filesystem = new Filesystem();
-
-        foreach($imagesNom as $nomImage){
-            $filesystem->remove($this->getParameter('images_directory').'/'.$nomImage['nom']);
-        }
-
-        $entityManager->remove($figure);
-        $entityManager->flush();
-
-        return $this->redirectToRoute('app_home');
     }
 }
